Added Opaque Keys Support
- Updated CSV names
- Updated URL regex
- Refactored view tests (using DDT)
