+++ .pep8
@@ -1,4 +1,4 @@
 [pep8]
 ignore=E501
 max_line_length=119
-exclude=settings,migrations
+exclude=settings,migrations,bower_components
