+++ analytics_dashboard/courses/views.py
@@ -2,6 +2,8 @@ import copy
 import datetime
 import json
 import logging
+import urllib
+from opaque_keys.edx.keys import CourseKey
 import requests
 
 from django.conf import settings
@@ -296,12 +298,22 @@ class EngagementTemplateView(CourseTemplateView):
 
 
 class CSVResponseMixin(object):
+    csv_filename_suffix = None
+
+    # pylint: disable=unused-argument
     def render_to_response(self, context, **response_kwargs):
-        response = HttpResponse(context['data'], content_type='text/csv',
-                                **response_kwargs)
-        response['Content-Disposition'] = 'attachment; filename="{0}"'.format(
-            context['filename'])
+        response = HttpResponse(self.get_data(), content_type='text/csv', **response_kwargs)
+        response['Content-Disposition'] = 'attachment; filename="{0}"'.format(self._get_filename())
         return response
+
+    def get_data(self):
+        raise NotImplementedError
+
+    def _get_filename(self):
+        course_key = CourseKey.from_string(self.course_id)
+        course_id = '-'.join([course_key.org, course_key.course, course_key.run])
+        filename = '{0}--{1}.csv'.format(course_id, self.csv_filename_suffix)
+        return urllib.quote(filename)
 
 
 class JSONResponseMixin(object):
@@ -416,41 +428,26 @@ class EngagementContentView(EngagementTemplateView):
 
 
 class CourseEnrollmentByCountryCSV(CSVResponseMixin, CourseView):
-    def get_context_data(self, **kwargs):
-        context = super(CourseEnrollmentByCountryCSV, self).get_context_data(**kwargs)
+    csv_filename_suffix = u'enrollment-location'
 
-        context.update({
-            'data': self.course.enrollment(demographic.LOCATION, data_format=data_format.CSV),
-            'filename': '{0}_enrollment_by_country.csv'.format(self.course_id)
-        })
-
-        return context
+    def get_data(self):
+        return self.course.enrollment(demographic.LOCATION, data_format=data_format.CSV)
 
 
 class CourseEnrollmentCSV(CSVResponseMixin, CourseView):
-    def get_context_data(self, **kwargs):
-        context = super(CourseEnrollmentCSV, self).get_context_data(**kwargs)
+    csv_filename_suffix = u'enrollment'
+
+    def get_data(self):
         end_date = datetime.datetime.utcnow().strftime(Client.DATE_FORMAT)
-
-        context.update({
-            'data': self.course.enrollment(data_format=data_format.CSV, end_date=end_date),
-            'filename': '{0}_enrollment.csv'.format(self.course_id)
-        })
-
-        return context
+        return self.course.enrollment(data_format=data_format.CSV, end_date=end_date)
 
 
 class CourseEngagementActivityTrendCSV(CSVResponseMixin, CourseView):
-    def get_context_data(self, **kwargs):
-        context = super(CourseEngagementActivityTrendCSV, self).get_context_data(**kwargs)
+    csv_filename_suffix = u'engagement-activity'
+
+    def get_data(self):
         end_date = datetime.datetime.utcnow().strftime(Client.DATE_FORMAT)
-
-        context.update({
-            'data': self.course.activity(data_format=data_format.CSV, end_date=end_date),
-            'filename': '{0}_engagement_activity_trend.csv'.format(self.course_id)
-        })
-
-        return context
+        return self.course.activity(data_format=data_format.CSV, end_date=end_date)
 
 
 class CourseHome(LoginRequiredMixin, RedirectView):
