+++ requirements/test.txt
@@ -17,3 +17,4 @@ testfixtures==4.0.1
 nose-ignore-docstring==0.2
 pyquery>=1.2.9
 selenium>=2.43.0
+ddt==0.8.0
